#17 Fixed a bug where boolean expressions inside trigger parameters where not parsed
